experiments/pyproject.toml
[project]
name = "experiments"
version = "0.1.0"
description = "Pinglab experiments"
requires-python = ">=3.10"
dependencies = [
    "pinglab",
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/experiments"]
